Program to read string from a user, remove first and last character of input string using slicing.
This program will exchange the first letter of the sentence or word with the last letter of word or sentence. If the user doen't enter any letter then it will print"String is empty", and if the user will give only one letter of input then it will print that same letter.